4.12.2024

+++ Assets/Scripts/SO/UnitSO.cs
@@ -4,7 +4,7 @@ using UnityEngine;
 public class UnitSO : ScriptableObject
 {
     [field:SerializeField] public float AttackRange { get; private set; }
-    [field:SerializeField] public int AttackDamage { get; private set; }
+    [field:SerializeField] public float AttackDamage { get; private set; }
     [field:SerializeField] public float TimeBetweenAttack { get; private set; }
     [field:SerializeField] public int MaxHealth { get; private set; }
     [field:SerializeField] public int Armor { get; private set; }
